seperate helper functions in utils
